Updated to 4.0 and adjusted project structure to be inline with godot standards
