+++ .gitignore
@@ -6,3 +6,6 @@
 .idea/
 demo/android
 tools/releases
+godot-android-share/build/
+godot-android-share/libs/godot-lib.4.0.3.stable.template_release.aar
+.gradle/
