+++ godot-android-share/src/main/AndroidManifest.xml
@@ -1,0 +1,11 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="org.godotengine.godot.plugin.GodotAndroidShare">
+
+    <application>
+
+        <meta-data
+            android:name="org.godotengine.plugin.v1.GodotAndroidShare"
+            android:value="org.godotengine.godot.plugin.GodotAndroidShare.GodotAndroidShare" />
+
+    </application>
+</manifest>
