+++ godot-android-share/src/main/java/org/godotengine/godot/plugin/GodotShare/GodotAndroidShare.java
@@ -1,0 +1,114 @@
+/*************************************************************************/
+/*  GodotAndroidShare.java                                                    */
+/*************************************************************************/
+/*                       This file is part of:                           */
+/*                           GODOT ENGINE                                */
+/*                      https://godotengine.org                          */
+/*************************************************************************/
+/* Copyright (c) 2007-2020 Juan Linietsky, Ariel Manzur.                 */
+/* Copyright (c) 2014-2020 Godot Engine contributors (cf. AUTHORS.md).   */
+/*                                                                       */
+/* Permission is hereby granted, free of charge, to any person obtaining */
+/* a copy of this software and associated documentation files (the       */
+/* "Software"), to deal in the Software without restriction, including   */
+/* without limitation the rights to use, copy, modify, merge, publish,   */
+/* distribute, sublicense, and/or sell copies of the Software, and to    */
+/* permit persons to whom the Software is furnished to do so, subject to */
+/* the following conditions:                                             */
+/*                                                                       */
+/* The above copyright notice and this permission notice shall be        */
+/* included in all copies or substantial portions of the Software.       */
+/*                                                                       */
+/* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,       */
+/* EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF    */
+/* MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.*/
+/* IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY  */
+/* CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,  */
+/* TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE     */
+/* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
+/*************************************************************************/
+
+package org.godotengine.godot.plugin.GodotAndroidShare;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.net.Uri;
+import android.util.Log;
+
+import org.godotengine.godot.Dictionary;
+import org.godotengine.godot.Godot;
+import org.godotengine.godot.plugin.GodotPlugin;
+import org.godotengine.godot.plugin.SignalInfo;
+import org.godotengine.godot.plugin.UsedByGodot;
+
+import androidx.annotation.NonNull;
+import androidx.collection.ArraySet;
+import androidx.core.content.FileProvider;
+
+import java.io.File;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Set;
+
+public class GodotAndroidShare extends GodotPlugin {
+
+	private static final String TAG = "godot";
+    private Activity activity;
+
+    public GodotAndroidShare(Godot godot) {
+        super(godot);
+        activity = godot.getActivity();
+    }
+
+    @UsedByGodot
+    public void shareText(String title, String subject, String text) {
+        Log.d(TAG, "shareText called");
+        Intent sharingIntent = new Intent(android.content.Intent.ACTION_SEND);
+        sharingIntent.setType("text/plain");
+        sharingIntent.putExtra(android.content.Intent.EXTRA_SUBJECT, subject);
+        sharingIntent.putExtra(android.content.Intent.EXTRA_TEXT, text);
+        activity.startActivity(Intent.createChooser(sharingIntent, title));
+    }
+
+    @UsedByGodot
+    public void sharePic(String path, String title, String subject, String text) {
+        Log.d(TAG, "sharePic called");
+
+        File f = new File(path);
+
+        Uri uri;
+        try {
+            uri = FileProvider.getUriForFile(activity, activity.getPackageName(), f);
+        } catch (IllegalArgumentException e) {
+            Log.e(TAG, "The selected file can't be shared: " + path);
+            return;
+        }
+
+        Intent shareIntent = new Intent(Intent.ACTION_SEND);
+        shareIntent.setType("image/*");
+        shareIntent.putExtra(Intent.EXTRA_SUBJECT, subject);
+        shareIntent.putExtra(Intent.EXTRA_TEXT, text);
+        shareIntent.putExtra(Intent.EXTRA_STREAM, uri);
+        shareIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+        activity.startActivity(Intent.createChooser(shareIntent, title));
+    }
+
+    @NonNull
+    @Override
+    public List<String> getPluginMethods() {
+        return Arrays.asList("sharePic", "shareText");
+    }
+
+    @NonNull
+    @Override
+    public String getPluginName() {
+        return "GodotAndroidShare";
+    }
+
+	@NonNull
+	@Override
+	public Set<SignalInfo> getPluginSignals() {
+		Set<SignalInfo> signals = new ArraySet<>();
+		return signals;
+	}
+}
