+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Sep 01 20:58:16 CEST 2022
+distributionBase=GRADLE_USER_HOME
+distributionUrl=https\://services.gradle.org/distributions/gradle-7.3.3-bin.zip
+distributionPath=wrapper/dists
+zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
