+++ local.properties
@@ -1,0 +1,8 @@
+## This file must *NOT* be checked into Version Control Systems,
+# as it contains information specific to your local configuration.
+#
+# Location of the SDK. This is only used by Gradle.
+# For customization when using a Version Control System, please read the
+# header note.
+#Sat May 27 09:16:07 EDT 2023
+sdk.dir=C\:\\Users\\finepointcgi\\AppData\\Local\\Android\\Sdk
