--- share-plugin/.gitignore
@@ -1,10 +1,0 @@
-*.iml
-.gradle
-local.properties
-.DS_Store
-/build
-/captures
-.externalNativeBuild
-.cxx
-/godot-lib.release/build
-/godot-lib.release/godot-lib.release.aar
