--- share-plugin/app/src/androidTest/java/shinnil/godot/plugin/android/godotshare/ExampleInstrumentedTest.java
@@ -1,27 +1,0 @@
-package shinnil.godot.plugin.android.godotshare;
-
-import android.content.Context;
-
-import androidx.test.platform.app.InstrumentationRegistry;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static org.junit.Assert.*;
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-@RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentedTest {
-    @Test
-    public void useAppContext() {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-
-        assertEquals("shinnil.godot.plugin.android.godotshare", appContext.getPackageName());
-    }
-}
