--- share-plugin/app/src/main/AndroidManifest.xml
@@ -1,11 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="shinnil.godot.plugin.android.godotshare">
-
-    <application
-        android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
-        android:supportsRtl="true"
-        android:theme="@style/AppTheme" />
-</manifest>
