--- share-plugin/build.gradle
@@ -1,32 +1,0 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-// TODO: Copy config.gradle from the build template of a specific Godot version.
-apply from: "../demo/android/build/config.gradle"
-apply from: 'version.gradle'
-
-buildscript {
-    apply from: "../demo/android/build/config.gradle"
-
-    repositories {
-        google()
-        mavenCentral()
-    }
-
-    dependencies {
-        classpath "com.android.tools.build:gradle:${versions.androidGradlePlugin}"
-
-
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
-    }
-}
-
-allprojects {
-    repositories {
-        google()
-        mavenCentral()
-    }
-}
-
-task clean(type: Delete) {
-    delete rootProject.buildDir
-}
