--- share-plugin/godot-lib.release/build.gradle
@@ -1,2 +1,0 @@
-configurations.maybeCreate("default")
-artifacts.add("default", file('godot-lib.release.aar'))
