--- share-plugin/godotshare/src/main/AndroidManifest.xml
@@ -1,21 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="shinnil.godot.plugin.android.godotshare">
-
-    <application>
-        <meta-data
-            android:name="org.godotengine.plugin.v1.GodotShare"
-            android:value="shinnil.godot.plugin.android.godotshare.GodotShare" />
-
-        <provider
-            android:name="androidx.core.content.FileProvider"
-            android:grantUriPermissions="true"
-            android:exported="false"
-            android:authorities="${applicationId}">
-
-            <meta-data
-                android:name="android.support.FILE_PROVIDER_PATHS"
-                android:resource="@xml/file_provider_paths"/>
-        </provider>
-    </application>
-</manifest>
-
