--- share-plugin/godotshare/src/main/res/xml/file_provider_paths.xml
@@ -1,3 +1,0 @@
-<paths>
-    <files-path path="/" name="filesRoot" />
-</paths>
