--- share-plugin/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Fri Apr 23 18:42:14 BRT 2021
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-7.3.3-bin.zip
