--- share-plugin/settings.gradle
@@ -1,3 +1,0 @@
-rootProject.name='GodotShare'
-include ':app', ':godot-lib.release'
-include ':godotshare'
