--- share-plugin/version.gradle
@@ -1,3 +1,0 @@
-ext.pluginVersionCode = 5
-ext.pluginVersionName = "4.0.0"
-ext.appcompatVersion = "1.5.1"
